// src/frontend/src/app/core/models/user.model.ts
export interface User {
  id: string;
  email: string;
  nome: string;
  userName: string;
  dataCriacao: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface SigninResponse {
  token: string;
  refreshToken: string;
  user: User;
}

export interface RegisterRequest {
  nome: string;
  userName: string;
  email: string;
  password: string;
}

export interface RefreshTokenRequest {
  refreshToken: string;
}

export interface RefreshTokenResponse {
  token: string;
  refreshToken: string;
}

export interface AuthState {
  isAuthenticated: boolean;
  user: User | null;
  token: string | null;
  refreshToken: string | null;
  isLoading: boolean;
  error: string | null;
}